Questionaria

+++ core/src/commonMain/kotlin/core/data/model/evaluation/EvaluationSettings.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class EvaluationSettings(
     val id: Int = 0,
     val measurement_repeat_period: String = "",
-    val measurement_repeat_times: Int = 0,
+    val measurement_repeat_times: Float = 0f,
     val measurement_begin_time: String = "",
     val measurement_last_time: String = "",
     val measurement_end_time: String? = "",
